add DI by Spring with JavaConfig
