+++ pom.xml
@@ -33,6 +33,11 @@
             <artifactId>maven-dependency-plugin</artifactId>
             <version>3.5.0</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-context</artifactId>
+            <version>5.2.24.RELEASE</version>
+        </dependency>
     </dependencies>
 
     <build>
